se agrega path absoluto, validaciones, try catch faltantes, campos category y status a los productos

+++ primera pre-entrega/src/routes/products.router.js
@@ -1,48 +1,86 @@
 import ProductManager from '../manager/ProductManager.js';
 import {Router} from 'express';
+import { __dirname } from "../utils.js";
+import path from 'node:path';
 
-const manager = new ProductManager('../src/files/productos.json');
+const productsFilePath = path.join(__dirname, "./files/products.json");
+const manager = new ProductManager(productsFilePath);
+
 const router = Router();
 
 //Traer todos los productos con query de limit
 router.get('/', async (req,res) => {
-    const products = await manager.getProducts();
-    const queryParamsLimit = parseInt(req.query.limit);
-    if(queryParamsLimit){
-        let result = products.slice(0, queryParamsLimit);
-        res.send(result);
-    }else{
-        res.send(products);
-    }
+    try {
+        const products = await manager.getProducts();
+        const queryParamsLimit = parseInt(req.query.limit);
+        if(queryParamsLimit){
+            let result = products.slice(0, queryParamsLimit);
+            res.send(result);
+        }else{
+            res.send(products);
+        }
+    } catch (error) {
+        res.status(400).send({ error: error.message })
+    }    
 });
 
 //Traer por id
 router.get('/:pid', async (req,res) => {
-    res.send(await manager.getProductById(req.params.pid));
+    try {
+        res.send(await manager.getProductById(req.params.pid));
+    } catch (error) {
+        res.status(400).send({ error: error.message })
+    }
 });
 
 //Agregar producto
 router.post('/', async (req, res) => {
-    const product = req.body; //Obtenerel json con los datos del producto
-    await manager.addProduct(product.title, product.description, product.price, product.thumbnail, product.stock);
-    res.send({status:'success', payload: product });
+    try {
+        const {title, description, price, thumbnail, stock, category} = req.body; //Obtenerel json con los datos del producto
+        if(!title || !description || !price ||  !thumbnail || !category || (!stock && !stock === 0)){
+            res.send({status:'not succes 400', payload: "faltan datos" })
+        } else {
+            const product = await manager.addProduct(title, description, price, thumbnail, stock, category);
+            res.send({status:'success', payload: product });
+        }
+        
+        
+    } catch (error) {
+        res.status(400).send({ error: error.message })
+    }
 });
 
 //Actualizar producto
 router.put('/:pid', async (req, res) => {
-    const product = req.body;
-    const id = parseInt(req.params.pid);
-    await manager.updateProduct(id, product);
-    res.send({status:'success', payload: {...product, id}});
+    try {
+        const id = parseInt(req.params.pid);
+        const prod = await manager.getProductById(id)
+        if(prod.id != id){
+            res.send({status:'not succes 400', payload: "id no fue encontrado" })
+        } else {
+            const product = req.body;
+            await manager.updateProduct(id, product);
+            res.send({status:'success', payload: {...product, id}});
+        }
+    } catch (error) {
+        res.status(400).send({ error: error.message })
+    }
 });
 
 //Borrar producto
 router.delete('/:pid', async (req, res) =>{
-    
-    const id = parseInt(req.params.pid);
-    const product = await manager.getProductById(id)
-    await manager.deleteProduct(id)
-    res.send({status:'success', payload: `Se elimino el producto: ${product.title} id:${product.id}` });
+    try {
+        const id = parseInt(req.params.pid);
+        const product = await manager.getProductById(id)
+        if(product.id != id){
+            res.send({status:'not succes 400', payload: "id no fue encontrado" })
+        } else {
+            await manager.deleteProduct(id)
+            res.send({status:'success', payload: `Se elimino el producto: ${product.title} id:${product.id}` });
+        }
+    } catch (error) {
+        res.status(400).send({ error: error.message })
+    }    
 })
 
 export default router;
